Add effective frequency
Like HWiNFO64, the popular monitoring tool for Windows, this patch
adds support for reading the cores' effective frequency or multiplier
by making use of the undocumented MSR, 0xC0010293.

This MSR was found by reverse-engineering Ryzen Master, the AMD software
for Windows overclocking and monitoring of Zen based processors. Although
this MSR is undocumented, since it is used in software written by AMD
themselves, it is safe to assume it is accurate.

The MSR returns two things, the FID or effective frequency ID and the FDID.
The FID, when divided by the FDID for a core, produces a frequency in
increments of 200 MHz -- the effective frequency.